Previous-prototypes/4.prototype-pandasai-openinterpreter/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.9
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.99.1
filelock==3.12.4
flatbuffers==23.5.26
fsspec==2023.9.2
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
monotonic==1.6
mpmath==1.3.0
numpy==1.26.0
onnxruntime==1.16.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pydantic==1.10.12
pydantic_core==2.10.0
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
